// js/stat-evaluation-player/src/generated/WallAerialShotStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type WallAerialShotStats = { count: number, high_confidence_count: number, is_last_wall_aerial_shot: boolean, last_wall_aerial_shot_time: number | null, last_wall_aerial_shot_frame: number | null, time_since_last_wall_aerial_shot: number | null, frames_since_last_wall_aerial_shot: number | null, last_confidence: number | null, best_confidence: number, cumulative_confidence: number, cumulative_takeoff_to_shot_time: number, cumulative_shot_height: number, };